requirements-dev.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
authlib==1.6.0; python_version >= '3.9'
bandit==1.8.6; python_version >= '3.9'
black==25.1.0; python_version >= '3.9'
certifi==2025.7.9; python_version >= '3.7'
cffi==1.17.1; platform_python_implementation != 'PyPy'
cfgv==3.4.0; python_version >= '3.8'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
coverage[toml]==7.9.2; python_version >= '3.9'
cryptography==45.0.5; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
distlib==0.3.9
dparse==0.6.4; python_version >= '3.7'
filelock==3.16.1; python_version >= '3.8'
flake8==7.3.0; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
identify==2.6.12; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
iniconfig==2.1.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
joblib==1.5.1; python_version >= '3.9'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==4.0.0; python_version >= '3.9'
mccabe==0.7.0; python_version >= '3.6'
mdurl==0.1.2; python_version >= '3.7'
mypy==1.16.1; python_version >= '3.9'
mypy-extensions==1.1.0; python_version >= '3.8'
nltk==3.9.1; python_version >= '3.8'
nodeenv==1.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==25.0; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
pbr==6.1.1; python_version >= '2.6'
platformdirs==4.3.8; python_version >= '3.9'
pluggy==1.6.0; python_version >= '3.9'
pre-commit==4.2.0; python_version >= '3.9'
psutil==6.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycodestyle==2.14.0; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pyflakes==3.4.0; python_version >= '3.9'
pygments==2.19.2; python_version >= '3.8'
pytest==8.4.1; python_version >= '3.9'
pytest-asyncio==1.0.0; python_version >= '3.9'
pytest-cov==6.2.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.4; python_version >= '3.8'
rich==14.0.0; python_full_version >= '3.8.0'
ruamel-yaml==0.18.14; python_version >= '3.8'
ruamel.yaml.clib==0.2.12; python_version < '3.14' and platform_python_implementation == 'CPython'
safety==3.5.2; python_version >= '3.8'
safety-schemas==0.0.14; python_version >= '3.7'
setuptools==80.9.0; python_version >= '3.9'
shellingham==1.5.4; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
stevedore==5.4.1; python_version >= '3.9'
tenacity==9.1.2; python_version >= '3.9'
tomlkit==0.13.3; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typer==0.16.0; python_version >= '3.7'
typing-extensions==4.14.1; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
virtualenv==20.31.2; python_version >= '3.8'
alembic==1.16.3; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
autoflake==2.3.1; python_version >= '3.8'
base58==2.1.1; python_version >= '3.5'
fastapi==0.116.0; python_version >= '3.8'
greenlet==3.2.3; python_version < '3.14' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
httptools==0.6.4; python_full_version >= '3.8.0'
isort==6.0.1; python_full_version >= '3.9.0'
jsonschema==4.24.0; python_version >= '3.9'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
mako==1.3.10; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pydantic-settings==2.10.1; python_version >= '3.9'
python-bitcoinrpc==1.0
python-dotenv==1.1.1; python_version >= '3.9'
redis==6.2.0; python_version >= '3.9'
referencing==0.36.2; python_version >= '3.9'
rpds-py==0.26.0; python_version >= '3.9'
sqlalchemy==2.0.41; python_version >= '3.7'
starlette==0.46.2; python_version >= '3.9'
structlog==25.4.0; python_version >= '3.8'
typing-inspection==0.4.1; python_version >= '3.9'
uvicorn[standard]==0.35.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.1.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
